refactoring: put production code into module utm_zone_info
