+++ utm_zone_info/__init__.py
@@ -1,4 +1,4 @@
-__version__ = 'v2.1.1'
+__version__ = 'v0.0.1'
 
 __all__ = [
     '__version__',
